% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temor.R
\name{OR_95CI}
\alias{OR_95CI}
\title{Inverse logistic link function}
\usage{
OR_95CI(coef, se, siglevel, roundto)
}
\arguments{
\item{x}{numeric vector}
}
\value{
OR Results
}
\description{
a function that returns the inverse of the logistic link function
}
\examples{
coef = c(1:3)
se = c(0.1,0.2,0.3)
OR_95CI(coef = coef, se = se, siglevel = 0.05, roundto=2)
}
\author{
Temoor Tayyab
}
