% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Example data}
\format{
A data frame with 100 rows and 3 variables.
\describe{
  \item{y}{binary outcome variable, 0s and 1s}
  \item{x1}{continuous explanatory variable}
  \item{x2}{binary explanatory variable}
}
}
\usage{
toydata
}
\description{
Toydata
}
\keyword{datasets}
